cppcheck report fixes

+++ jolt_os/syscore/ota.c
@@ -60,7 +60,7 @@ esp_err_t jolt_ota_ymodem( int8_t *progress )
     /***************************************
      * Receive and Write data to partition *
      ***************************************/
-    size_t binary_file_length;
+    int binary_file_length;
     binary_file_length = ymodem_receive_write( (void *)jolt_ota_handle, update_partition->size, NULL,
                                                &ota_ymodem_write_wrapper, progress );
     if( binary_file_length <= 0 ) {
